Add session store for account logins
A helper is added with a function getLogin($accessRight), which returns the logged in account based on the accessRight or nil if not logged in or accessRight does not match the account. $accessRight can be "user" or "staff" and will assume full access for logged in users if anything else is supplied.

Login username is stored in the session variable, use the User model to query for additional information.

For each page, attempt to query the login user based on the access rights of the page. One can redirect to /page/account/login.php if access right does not match. login.php automatically redirects to the corresponding landing pages if there is a logged in user.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>