Add show view for staff
Staff do not see fine warnings, but are allowed to edit answers should there be one. On top of that if a question has yet to be taken by any staff, a staff may choose to accept the question. Questions that the staff have accepted will show an answer button instead of the accept question button.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>